requirements.txt
ai2thor==2.1.0
-e git+https://github.com/alfworld/alfworld.git@c0501c807fa77dd25ecfc86e92cd782226f5c74f#egg=alfworld
backcall==0.2.0
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.2.2
configparser==5.0.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
datasets==1.5.0
decorator==4.4.2
dill==0.3.3
docker-pycreds==0.4.0
docopt==0.6.2
fast-downward @ https://github.com/MarcCote/downward/archive/faster_replan.zip
filelock==3.0.12
Flask==1.1.2
fsspec==0.9.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
gym==0.15.4
h5py==3.2.1
hashids==1.3.1
huggingface-hub==0.0.8
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
jericho==3.0.4
Jinja2==2.11.3
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mementos==1.3.1
more-itertools==8.7.0
multiprocess==0.70.11.1
murmurhash==1.0.5
networkx==2.5.1
nltk==3.6.1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
parso==0.8.2
pathtools==0.1.2
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
preshed==3.0.5
progressbar2==3.53.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycocotools==2.0.2
pycparser==2.20
pydantic==1.7.3
pyglet==1.3.2
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
revtok==0.0.3
sacremoses==0.0.44
scipy==1.6.2
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.5
spacy-legacy==3.0.2
srsly==2.4.1
subprocess32==3.5.4
TatSu==4.4.0
tensorboardX==1.8
textworld @ https://github.com/MarcCote/TextWorld/archive/handcoded_expert_integration.zip
thinc==8.0.2
tokenizers==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.5.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.1.8.9
vocab==0.0.5
wandb==0.10.26
wasabi==0.8.2
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
xxhash==2.0.2
yarg==0.1.9
